PromethION side changes to tar comamnds for alpta

+++ poreduck/promethion_alpha_light.py
@@ -24,6 +24,8 @@ def get_args():
     parser.add_argument("--flowcellID", help="Flowcell ID to add to suffix of tar file")
     parser.add_argument("--rnumber", help="Add random number to folder")
     parser.add_argument("--md5", help="File to append to md5sum")
+    parser.add_argument("--inplace", action='store_true', default=False, help="Remove folders as well")
+    parser.add_argument("--overwrite", action='store_true', default=False, help="Overwrite output file rather than append to it")
     parser.add_argument("--dry-run", dest='dry_run', action='store_true', default=False,
                         help="Don't actually tar anything, just output the logs")
     args = parser.parse_args()
@@ -54,13 +56,23 @@ def get_output_path(folder_path, output_name):
     return output_path
 
 
-def tar_up_folder(folder_path, output_path, dry_run=False):
+def tar_up_folder(folder_path, output_path, overwrite=False, inplace=False, dry_run=False):
     # Tar up the folder provided
     # Get the output path
     logging.info("Output path is %s" % output_path)
 
     # Normalise the output path, it's come straight from args
     folder_path = os.path.normpath(folder_path)
+
+    # To  overwrite or not to overwrite
+    if overwrite:
+        file_handler_setting = "w|gz"
+    else:
+        # Can't actually append a compressed tar yet We're going to sys.exit  if the file exist
+        if os.path.isfile(output_path):
+            sys.exit("Tar exists, not overwriting")
+        file_handler_setting = "w|gz"  # Default is append 
+	
 
     # Get number of files in the path
     fast5_files = [fast5_file
@@ -76,17 +88,23 @@ def tar_up_folder(folder_path, output_path, dry_run=False):
         start_time = datetime.now()
         logging.info("Starting tarring %s into %s" % (folder_path, output_path))
         logging.info("%d files to tar" % len(fast5_files))
+
         # Open up the output_path file
-        archive = tarfile.open(output_path, "w|gz")
+        archive = tarfile.open(output_path, file_handler_setting)
         # Add each of the fast5 files to the archive
         for fast5_file in fast5_files:
             input_file = os.path.join(folder_path, fast5_file)
             output_file = os.path.join(os.path.basename(folder_path), fast5_file)
             # Add file to archive
             archive.add(input_file, arcname=output_file)
-            # Remove file from filesystem
-            os.remove(input_file)
+            # If inplace also remove the input file from the system
+            if inplace:
+                # Remove file from filesystem
+                os.remove(input_file)
+        # Close the archive
         archive.close()
+        
+        # Log the time taken to write the archive file
         end_time = datetime.now()
         diff_time = end_time - start_time
         logging.info("Finished tarring %s in %s" % (folder_path, output_path))
@@ -98,11 +116,14 @@ def tar_up_folder(folder_path, output_path, dry_run=False):
 
 def get_md5sum(output_path):
     logging.info("Obtaining the md5sum for %s" % output_path)
+
     # Grab the md5sum of the file
     md5_command = ['md5sum', output_path]
     md5_proc = subprocess.run(md5_command, stdout=subprocess.PIPE)
     md5_output = md5_proc.stdout.decode().splitlines()[0]
     logging.info("Obtained %s as md5 for %s" % (md5_output, output_path))
+
+    # Return the md5 of the file for writing to a checksums file
     return md5_output
 
 
@@ -119,7 +140,7 @@ def main():
     output_name = get_output_name(args)
     output_path = get_output_path(args.folder_path, output_name)
     # Tar up folder
-    tar_up_folder(args.folder_path, output_path, args.dry_run)
+    tar_up_folder(args.folder_path, output_path, args.overwrite, args.inplace, args.dry_run)
     # Get md5
     if not args.dry_run:
         md5sum = get_md5sum(output_path)
